restructure: extract JSON integrations into dedicated modules, fold the OpenAPI auto-config into the swagger module
The core typed-ids artifact no longer bundles serialization integrations: Jackson 2 moves to typed-ids-jackson2, Gson to typed-ids-gson, kotlinx to typed-ids-kotlinx-serialization (packages unchanged - consumers only add a dependency). The Spring Boot auto-config moves into typed-ids-openapi-swagger-jakarta, which now works standalone via the swagger SPI and auto-configures under Boot; the typed-ids-openapi-springdoc artifact ceases to exist (it consumed zero springdoc API). Module floors follow what the oldest supported Spring Boot (3.0.0) manages. The README migration table is the 0.12 upgrade guide.
