+++ gradle/libs.versions.toml
@@ -5,11 +5,13 @@ junit-jupiter = "5.10.3"
 jackson = "2.18.1"
 testcontainers = "1.20.3"
 springBoot = "3.5.8"
+springBoot30 = "3.0.0"
 
 [libraries]
 uuidGenerator = { module = "com.fasterxml.uuid:java-uuid-generator", version = "5.1.1" }
 jetbrains-annotations = { module = "org.jetbrains:annotations", version = "26.0.2" }
 jackson-databind = { group = "com.fasterxml.jackson.core", name = "jackson-databind", version.ref = "jackson" }
+jackson-databind-v2141 = { group = "com.fasterxml.jackson.core", name = "jackson-databind", version = "2.14.1" }
 jackson-annotations = { group = "com.fasterxml.jackson.core", name = "jackson-annotations", version.ref = "jackson" }
 gson = { module = "com.google.code.gson:gson", version = "2.10.1" }
 autoService-processor = { group = "com.google.auto.service", name = "auto-service", version.ref = "googleAutoService" }
@@ -46,5 +48,6 @@ hypersistence-utils-hibernate71 = { group = "io.hypersistence", name = "hypersis
 hypersistence-tsid = { module = "io.hypersistence:hypersistence-tsid", version = "2.1.4" }
 kotlinx-serialization = { module = "org.jetbrains.kotlinx:kotlinx-serialization-json", version = "1.7.0" }
 swagger-v3-core-jakarta = { module = "io.swagger.core.v3:swagger-core-jakarta", version = "2.2.35" }
-springdoc-openapi-starter-common = { module = "org.springdoc:springdoc-openapi-starter-common", version = "2.1.0" }
-springdoc-openapi-spring-configuration-processor = { module = "org.springframework.boot:spring-boot-configuration-processor", version.ref = "springBoot" }
+spring-boot-v300 = { group = "org.springframework.boot", name = "spring-boot", version.ref = "springBoot30" }
+spring-boot-autoconfigure-v300 = { group = "org.springframework.boot", name = "spring-boot-autoconfigure", version.ref = "springBoot30" }
+spring-boot-configuration-processor = { group = "org.springframework.boot", name = "spring-boot-configuration-processor", version.ref = "springBoot30" }
