+++ modules/typed-ids-gson/build.gradle.kts
@@ -1,0 +1,21 @@
+plugins {
+    id("framefork.java-public")
+}
+
+dependencies {
+    api(project(":typed-ids"))
+
+    compileOnly(libs.gson)
+
+    compileOnly(libs.jetbrains.annotations)
+
+    compileOnly(libs.autoService.annotations)
+    annotationProcessor(libs.autoService.processor)
+
+    testImplementation(project(":typed-ids-testing"))
+    testImplementation(testFixtures(project(":typed-ids")))
+    testImplementation(libs.gson)
+    testRuntimeOnly("org.junit.platform:junit-platform-launcher")
+}
+
+project.description = "TypeIds (de)serialization with Gson"
