+++ modules/typed-ids-gson/src/main/java/org/framefork/typedIds/uuid/json/gson/ObjectUuidTypeAdapterFactory.java
