+++ modules/typed-ids-gson/src/test/java/org/framefork/typedIds/uuid/json/gson/ObjectUuidGsonTest.java
