+++ modules/typed-ids-jackson2/src/main/java/org/framefork/typedIds/bigint/json/jackson/ObjectBigIntIdJacksonModule.java
