+++ modules/typed-ids-jackson2/src/test/java/org/framefork/typedIds/uuid/json/jackson/ObjectUuidJacksonModuleTest.java
