+++ modules/typed-ids-kotlinx-serialization/build.gradle.kts
@@ -1,0 +1,13 @@
+plugins {
+    id("framefork.java-public")
+}
+
+dependencies {
+    api(project(":typed-ids"))
+
+    compileOnly(libs.kotlinx.serialization)
+
+    compileOnly(libs.jetbrains.annotations)
+}
+
+project.description = "TypeIds (de)serialization with Kotlin Serialization"
