+++ modules/typed-ids-kotlinx-serialization/src/main/kotlin/org/framefork/typedIds/uuid/json/kotlinxSerialization/ObjectUuidKotlinxSerializationModule.kt
