+++ modules/typed-ids-openapi-swagger-jakarta/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
