+++ modules/typed-ids/build.gradle.kts
@@ -1,5 +1,6 @@
 plugins {
     id("framefork.java-public")
+    id("java-test-fixtures")
 }
 
 dependencies {
@@ -10,17 +11,7 @@ dependencies {
 
     compileOnly(libs.jetbrains.annotations)
 
-    compileOnly(libs.autoService.annotations)
-    annotationProcessor(libs.autoService.processor)
-
-    compileOnly(libs.jackson.databind)
-    compileOnly(libs.gson)
-    compileOnly(libs.kotlinx.serialization)
-
     testImplementation(project(":typed-ids-testing"))
-    testImplementation(libs.jackson.databind)
-    testImplementation(libs.gson)
-    testImplementation(libs.kotlinx.serialization)
     testRuntimeOnly("org.junit.platform:junit-platform-launcher")
 }
