+++ modules/typed-ids/src/testFixtures/java/org/framefork/typedIds/bigint/ObjectBigIntIdMock.java
