+++ modules/typed-ids/src/testFixtures/java/org/framefork/typedIds/uuid/ObjectUuidMock.java
