+++ testing/testing-typed-ids-springdoc-2x-openapi/build.gradle.kts
@@ -7,7 +7,8 @@ plugins {
 
 dependencies {
     implementation(project(":typed-ids"))
-    implementation(project(":typed-ids-openapi-springdoc"))
+    implementation(project(":typed-ids-jackson2"))
+    implementation(project(":typed-ids-openapi-swagger-jakarta"))
 
     api(platform("org.springframework.boot:spring-boot-dependencies:3.5.8"))
     api("org.springframework.boot:spring-boot-starter-web")
@@ -42,5 +43,5 @@ tasks.named("build") {
 }
 
 tasks.named("forkedSpringBootRun") {
-    dependsOn(":typed-ids-openapi-springdoc:jar", ":typed-ids-openapi-swagger-jakarta:jar")
+    dependsOn(":typed-ids-jackson2:jar", ":typed-ids-openapi-swagger-jakarta:jar")
 }
